handle reruns and diff OSes

+++ init/setup_ubuntu.sh
@@ -9,24 +9,28 @@ apt update -y
 apt upgrade -y
 
 ## 1. Setup tailscale
-curl -fsSL https://tailscale.com/install.sh | sh
+if [ ! -f "/usr/bin/tailscale" ]; then
+    curl -fsSL https://tailscale.com/install.sh | sh
+fi
 tailscale up --ssh
 
 ## 2. Install docker
-# Add Docker's official GPG key:
-apt-get install -y ca-certificates curl
-install -m 0755 -d /etc/apt/keyrings
-curl -fsSL https://download.docker.com/linux/ubuntu/gpg -o /etc/apt/keyrings/docker.asc
-chmod a+r /etc/apt/keyrings/docker.asc
+if [ ! -f "/usr/bin/docker" ]; then
+    # Add Docker's official GPG key:
+    apt-get install -y ca-certificates curl
+    install -m 0755 -d /etc/apt/keyrings
+    curl -fsSL https://download.docker.com/linux/ubuntu/gpg -o /etc/apt/keyrings/docker.asc
+    chmod a+r /etc/apt/keyrings/docker.asc
 
-# Add the repository to Apt sources:
-echo \
-  "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/ubuntu \
-  $(. /etc/os-release && echo "$VERSION_CODENAME") stable" | \
-  tee /etc/apt/sources.list.d/docker.list > /dev/null
-apt-get update -y
+    # Add the repository to Apt sources:
+    echo \
+        "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/ubuntu \
+        $(. /etc/os-release && echo "$VERSION_CODENAME") stable" | \
+        tee /etc/apt/sources.list.d/docker.list > /dev/null
+    apt-get update -y
 
-apt-get install -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
+    apt-get install -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
+fi
 
 ## 3. Install mahogany
 wget https://github.com/mpoegel/mahogany/releases/download/latest/mahogany_Linux_x86_64.tar.gz
